Test Coverage for New Version #13
- Migrated old tests over to new version
- Removed old src and test code
