+++ src/Container/ContainerDecorator.php
@@ -4,6 +4,7 @@ namespace Krak\Cargo\Container;
 
 use Krak\Cargo;
 
+/** Stores values for the containers */
 abstract class ContainerDecorator extends AbstractContainer
 {
     protected $container;
@@ -16,11 +17,8 @@ abstract class ContainerDecorator extends AbstractContainer
         $this->container = clone $this->container;
     }
 
-    public function get($id, Cargo\Container $container = null) {
-        return $this->container->get($id, $container ?: $this);
-    }
-    public function has($id) {
-        return $this->container->has($id);
+    public function make($id, array $params = [], Cargo\Container $c = null) {
+        return $this->container->make($id, $params, $c ?: $this);
     }
     public function remove($id) {
         return $this->container->remove($id);
@@ -28,10 +26,16 @@ abstract class ContainerDecorator extends AbstractContainer
     public function add($id, $box, array $opts = []) {
         return $this->container->add($id, $box, $opts);
     }
+    public function has($id) {
+        return $this->container->has($id);
+    }
     public function box($id) {
         return $this->container->box($id);
     }
     public function keys() {
         return $this->container->keys();
+    }
+    public function count() {
+        return $this->container->count();
     }
 }
