--- src/Container/CountableTrait.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Krak\Cargo\Container;
-
-trait CountableTrait
-{
-    public function count() {
-        return count($this->keys());
-    }
-}
