+++ src/Exception/ContainerException.php
@@ -3,8 +3,9 @@
 namespace Krak\Cargo\Exception;
 
 use Psr\Container;
+use RuntimeException;
 
-class ContainerException extends \RuntimeException implements Container\ContainerExceptionInterface
+class ContainerException extends RuntimeException implements Container\ContainerExceptionInterface
 {
 
 }
