+++ src/Unbox/EnvUnbox.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace Krak\CargoV2\Unbox;
+namespace Krak\Cargo\Unbox;
 
-use Krak\CargoV2;
+use Krak\Cargo;
 
-class EnvUnbox implements CargoV2\Unbox
+class EnvUnbox implements Cargo\Unbox
 {
     private $unbox;
     private $getenv;
 
-    public function __construct(CargoV2\Unbox $unbox, $getenv = 'getenv') {
+    public function __construct(Cargo\Unbox $unbox, $getenv = 'getenv') {
         $this->unbox = $unbox;
         $this->getenv = $getenv;
     }
 
-    public function unbox($box, CargoV2\Container $container, array $params) {
+    public function unbox($box, Cargo\Container $container, array $params) {
         list($value, $opts) = $box;
-        if (!CargoV2\Container\optsEnv($opts)) {
+        if (!Cargo\Container\optsEnv($opts)) {
             return $this->unbox->unbox($box, $container, $opts);
         }
