+++ src/Unbox/ServiceUnbox.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace Krak\CargoV2\Unbox;
+namespace Krak\Cargo\Unbox;
 
 use Closure;
-use Krak\CargoV2;
+use Krak\Cargo;
 
-class ServiceUnbox implements CargoV2\Unbox
+class ServiceUnbox implements Cargo\Unbox
 {
-    public function unbox($box, CargoV2\Container $container, array $params) {
+    public function unbox($box, Cargo\Container $container, array $params) {
         list($value, $opts) = $box;
 
-        if (CargoV2\Container\optsAlias($opts)) {
+        if (Cargo\Container\optsAlias($opts)) {
             return $container->make($value, $params);
         }
 
-        if (!$value instanceof Closure) {
+        if (Cargo\Container\optsValue($opts) || !$value instanceof Closure) {
             return $value;
         }
 
-        $wrapped = CargoV2\Container\optsWrapped($opts);
+        $wrapped = Cargo\Container\optsWrapped($opts);
         if ($wrapped) {
             $unboxed_value = $this->unbox($wrapped, $container, $params);
             return $value($unboxed_value, $container, $params);
