--- srcv2/Exception/ContainerException.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Krak\CargoV2\Exception;
-
-use Psr\Container;
-use RuntimeException;
-
-class ContainerException extends RuntimeException implements Container\ContainerExceptionInterface
-{
-
-}
