--- srcv2/Exception/NotFoundException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Krak\CargoV2\Exception;
-
-use Psr\Container;
-
-class NotFoundException extends ContainerException implements Container\NotFoundExceptionInterface
-{
-
-}
