--- srcv2/ServiceProvider.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Krak\CargoV2;
-
-interface ServiceProvider
-{
-    public function register(Container $container);
-}
