+++ test/detect-cycles.spec.php
